Add user viewmodel with updated diagrams (#12)
