+++ SchoolManagerViewModel/AddUserViewModel.cs
@@ -1,112 +1,30 @@
 ﻿using SchoolManagerModel.Entities;
-using SchoolManagerModel.Entities.UserModel;
 using SchoolManagerModel.Managers;
 using SchoolManagerModel.Persistence;
-using SchoolManagerModel.Utils;
-using SchoolManagerModel.Validators;
 using SchoolManagerViewModel.Commands;
+using SchoolManagerViewModel.EntityViewModels;
 using System.Collections.ObjectModel;
-using System.Resources;
 
 namespace SchoolManagerViewModel;
 
 public class AddUserViewModel : ViewModelBase
 {
     #region Private Fields
-    private string _username = string.Empty;
-    private string _password = string.Empty;
-    private string _confirmPassword = string.Empty;
-    private string _email = string.Empty;
-    private string _firstName = string.Empty;
-    private string _lastName = string.Empty;
+
     private bool _isStudent = true;
     private bool _resetPassword = false;
     private bool _allSubjectsSelected = false;
-    private readonly ResourceManager _resourceManager;
     private Role _selectedRole = Role.Student;
     private Class? _class;
     private ObservableCollection<Class> _classes = [];
     private ObservableCollection<CheckBoxListItem<Subject>> _subjects = [];
-
-    private string _usernameValidatorErrors = string.Empty;
-    private string _emailValidatorErrors = string.Empty;
-    private string _passwordValidatorErrors = string.Empty;
-    private string _confirmPasswordValidatorErrors = string.Empty;
-    private string _firstNameValidatorErrors = string.Empty;
-    private string _lastNameValidatorErrors = string.Empty;
     private bool _hasSelectedClass;
+
     #endregion
 
     #region Public Properties
-    public string Username
-    {
-        get => _username;
-        set
-        {
-            SetField(ref _username, value, nameof(Username));
-            ValidateUsernameAsync(_username);
-            AddUserCommand.NotifyCanExecuteChanged();
-        }
-    }
 
-    public string Password
-    {
-        get => _password;
-        set
-        {
-            SetField(ref _password, value, nameof(Password));
-            var passwordValidator = new PasswordValidator(_resourceManager).Validate(_password);
-            PasswordValidatorErrors = ValidationErrors.GetErrorsFormatted(passwordValidator);
-            AddUserCommand.NotifyCanExecuteChanged();
-        }
-    }
-
-    public string ConfirmPassword
-    {
-        get => _confirmPassword;
-        set
-        {
-            SetField(ref _confirmPassword, value, nameof(ConfirmPassword));
-            var confirmPasswordValidator = new ConfirmPasswordValidator(_resourceManager).Validate((_password, _confirmPassword));
-            ConfirmPasswordValidatorErrors = ValidationErrors.GetErrorsFormatted(confirmPasswordValidator);
-            AddUserCommand.NotifyCanExecuteChanged();
-        }
-    }
-
-    public string Email
-    {
-        get => _email;
-        set
-        {
-            SetField(ref _email, value, nameof(Email));
-            ValidateEmailAsync(_email);
-            AddUserCommand.NotifyCanExecuteChanged();
-        }
-    }
-
-    public string FirstName
-    {
-        get => _firstName;
-        set
-        {
-            SetField(ref _firstName, value, nameof(FirstName));
-            var validator = new NotEmptyValidator(_resourceManager);
-            FirstNameValidatorErrors = ValidationErrors.GetErrorsFormatted(validator.Validate(_firstName));
-            AddUserCommand.NotifyCanExecuteChanged();
-        }
-    }
-
-    public string LastName
-    {
-        get => _lastName;
-        set
-        {
-            SetField(ref _lastName, value, nameof(LastName));
-            var validator = new NotEmptyValidator(_resourceManager);
-            LastNameValidatorErrors = ValidationErrors.GetErrorsFormatted(validator.Validate(_lastName));
-            AddUserCommand.NotifyCanExecuteChanged();
-        }
-    }
+    public ValidatedUserViewModel User { get; }
 
     public Role SelectedRole
     {
@@ -151,42 +69,6 @@ public class AddUserViewModel : ViewModelBase
         }
     }
 
-    public string UsernameValidatorErrors
-    {
-        get => _usernameValidatorErrors;
-        set => SetField(ref _usernameValidatorErrors, value, nameof(UsernameValidatorErrors));
-    }
-
-    public string EmailValidatorErrors
-    {
-        get => _emailValidatorErrors;
-        set => SetField(ref _emailValidatorErrors, value, nameof(EmailValidatorErrors));
-    }
-
-    public string PasswordValidatorErrors
-    {
-        get => _passwordValidatorErrors;
-        set => SetField(ref _passwordValidatorErrors, value, nameof(PasswordValidatorErrors));
-    }
-
-    public string ConfirmPasswordValidatorErrors
-    {
-        get => _confirmPasswordValidatorErrors;
-        set => SetField(ref _confirmPasswordValidatorErrors, value, nameof(ConfirmPasswordValidatorErrors));
-    }
-
-    public string FirstNameValidatorErrors
-    {
-        get => _firstNameValidatorErrors;
-        set => SetField(ref _firstNameValidatorErrors, value, nameof(FirstNameValidatorErrors));
-    }
-
-    public string LastNameValidatorErrors
-    {
-        get => _lastNameValidatorErrors;
-        set => SetField(ref _lastNameValidatorErrors, value, nameof(LastNameValidatorErrors));
-    }
-
     public ObservableCollection<CheckBoxListItem<Subject>> Subjects
     {
         get => _subjects;
@@ -223,10 +105,11 @@ public class AddUserViewModel : ViewModelBase
     #endregion
 
     #region Constructor
-    public AddUserViewModel(ResourceManager resourceManager)
+    public AddUserViewModel()
     {
-        _resourceManager = resourceManager;
+        User = new ValidatedUserViewModel();
         AddUserCommand = new AddUserCommand(this);
+        User.PropertyChanged += (_, _) => AddUserCommand.NotifyCanExecuteChanged();
         SelectAllSubjectsCommand = new SelectAllSubjectsCommand(this);
         GetClassesAsync();
         Class = _classes.FirstOrDefault();
@@ -251,42 +134,6 @@ public class AddUserViewModel : ViewModelBase
         foreach (var subject in subjects)
         {
             Subjects.Add(new CheckBoxListItem<Subject> { Item = subject });
-        }
-    }
-
-    private async void ValidateUsernameAsync(string username)
-    {
-        try
-        {
-            using var dbContext = new SchoolDbContext();
-            var userDatabase = new UserDatabase(dbContext);
-            var userManager = new UserManager(userDatabase);
-            var validator = new ValidNotExistsUserValidator(userManager, _resourceManager);
-
-            var validationResult = await validator.ValidateAsync(new User { Username = username });
-            UsernameValidatorErrors = ValidationErrors.GetErrorsFormatted(validationResult);
-        }
-        catch (Exception ex)
-        {
-            UsernameValidatorErrors = ex.Message;
-        }
-    }
-
-    private async void ValidateEmailAsync(string email)
-    {
-        try
-        {
-            using var dbContext = new SchoolDbContext();
-            var userDatabase = new UserDatabase(dbContext);
-            var userManager = new UserManager(userDatabase);
-            var validator = new ValidNotExistsEmailValidator(userManager, _resourceManager);
-
-            var validationResult = await validator.ValidateAsync(new User { Email = email });
-            EmailValidatorErrors = ValidationErrors.GetErrorsFormatted(validationResult);
-        }
-        catch (Exception ex)
-        {
-            EmailValidatorErrors = ex.Message;
         }
     }
